web28-googleMaps: final logic moved to google maps index
